scripts/intelligence/requirements.txt
ddgs>=7.0.0
pyyaml>=6.0
requests>=2.31.0
feedparser>=6.0.10
arxiv>=2.1.0
python-dateutil>=2.8.2
